Front-end organization and CSV formattring (new) split batch generation into own component/file (new) Admin tools now supports multiple collapsible areas (new) server-side routines for CSV formatting of asset results (new) convertFileSizeToString helper added making sizes human readable (fix) consolidated styles removing obsolete
